[MODE: RESEARCH]
分析这个 git diff，我观察到以下变化：

1. **README 文档更新**：
   - 修改了管理员账户初始化的描述（从"Admin Account"改为"super administrator account"）
   - 更新了默认凭据（从 `admin123` 改为 `Admin@123456`）
   - 调整了文档格式（缩进和空行）
   - 中文版 README 也进行了相应更新，并修改了初始化命令（从 `init_admin.bat` 改为 `python init_admin.py`）

2. **删除文件**：
   - 删除了 `init_admin.

+++ src/web/models/user.py
@@ -44,9 +44,9 @@ class User(Base):
     """
     __tablename__ = "users"
 
-    # 主键 - 使用BIGINT自增ID，与业务无关
+    # 主键 - 使用INT自增ID，与业务无关
     user_id = Column(
-        BigInteger, 
+        Integer, 
         primary_key=True, 
         index=True, 
         autoincrement=True,
